requirements.txt
streamlit
pandas
streamlit-autorefresh
os
datetime
dateutil.tz
streamlit_autorefresh
plotly.express
streamlit.components.v1